Implement target backend for AIRBIN
AIRBIN is a stand-alone file format that contains a series of configuration
instructions for AIE. This new backend outputs an ELF-compatible file that
contains all of the register values needed for configuration. The file is
uploaded to the control processor that executes the series of commands,
instead of the host executing each command one by one.

This depends on the 'elfutils' library.
https://github.com/jnider/elfutils.git

Configure elfutils with:
./configure --program-prefix="air-" --disable-debuginfod

Generate an AIRBIN by passing "--aie-generate-airbin" to aircc.py.

Signed-off-by: Joel Nider <joel.nider@amd.com>